+++ PUT ME IN A CLIENT PATCH/Interface/D3_Parangon/Icons/_D3mantraofretribution.blp
