+++ PUT ME IN A CLIENT PATCH/Interface/D3_Parangon/Icons/garr_currencyicon-xp.blp
