+++ PUT ME IN A CLIENT PATCH/Interface/D3_Parangon/buttons/ui-dialogbox-button-gold-up.blp
