+++ PUT ME IN A CLIENT PATCH/Interface/D3_Parangon/parangon/ButtonBorder_Hover.blp
