+++ PUT ME IN A CLIENT PATCH/Interface/D3_Parangon/parangon/LargeButtonBorder_Hover.blp
