+++ PUT ME IN A CLIENT PATCH/Interface/D3_Parangon/parangon/parangon_frame.blp
