--- Patch-Z.MPQ/Interface/AddOns/AIO_Client/AIO_Client.toc
@@ -1,16 +1,0 @@
-## Interface: 30300
-## Title: AIO_Client
-## Notes: Allows sending addons from server to client and allows client-server communication.
-## Author: Rochet2 <https://github.com/Rochet2>
-## SavedVariables: AIO_sv, AIO_sv_Addons
-## SavedVariablesPerCharacter: AIO_sv_char
-
-#dependencies
-Dep_LibWindow-1.1\LibStub.lua
-Dep_LibWindow-1.1\LibWindow-1.1\LibWindow-1.1.lua
-Dep_Smallfolk\smallfolk.lua
-lualzw-zeros\lualzw.lua
-queue.lua
-
-#core
-AIO.lua
