--- Patch-Z.MPQ/Interface/AddOns/AIO_Client/Dep_LibWindow-1.1/Changelog-LibWindow-1.1-r12.txt
@@ -1,13 +1,0 @@
-------------------------------------------------------------------------
-r12 | nevcairiel | 2014-10-14 19:12:48 +0000 (Tue, 14 Oct 2014) | 1 line
-Changed paths:
-   M /trunk/LibWindow-1.1.toc
-
-Update TOC
-------------------------------------------------------------------------
-r11 | mikk | 2013-03-10 14:27:14 +0000 (Sun, 10 Mar 2013) | 1 line
-Changed paths:
-   M /trunk/LibWindow-1.1.toc
-
-TOC 50200
-------------------------------------------------------------------------
